requirements.txt
# Discord library for bot functionality
discord.py>=2.3.0,<3.0.0

# Environment variable management
python-dotenv>=1.0.0,<2.0.0

# System information and process monitoring
psutil>=5.9.0,<6.0.0

# Image processing for UFO effects
Pillow>=10.0.0,<11.0.0
aiohttp>=3.8.0,<4.0.0

# Google Gemini AI for alien chat
google-generativeai>=0.3.0

# Web dashboard for monitoring
Flask>=3.0.0,<4.0.0